pothole detection - MLP
i saw that it was probably not as good an idea as i once thought to use LSTM to solve this problem.
ISSUE: dataset. There were too many negatives for holes compared to positives, so that, even if the model just guessed negative every time, it would still score 90+% accuracy.
SOLUTION: transform the dataset into images. These images are not so big, if you really were to go into the comparison they would just be (this was the best configuration i found) 10x10 pixels, so that the rate of images that contain potholes to images that don't is 50%. These images store sensor information and are fed straight into 2 fully connected layer, and they return a single value, either 0 or 1, '1' if the image has a pothole, 0 if it does not.

TODO: test accuracy on validation set